[FIX] mrp_account_enterprise: remove cost_analysis duplicate lines
Steps to reproduce :
- Create a Manufacturing Order with an operation
- Start the operation with the current user
- Log in another user and start the operation with him as well
- Wait a little and finish the operation and close the MO
- On the Cost Analysis report will be two line for this workcenter +
operation, both containing the full duration of the operation.

Following odoo/enterprise#23032, employee can be directly assigned to a
workorder. This changed the purpose of the Cost Analysis report a bit,
since now it shows operation lines for the workcenter AND lines for
their employee, to separate machine cost & labour cost.

Since the purpose of this query is now to have the machine cost, there
is little point of keeping the `group by` on the users in the query.
This will result on having only a single ligne per (workcenter,
operation).

closes odoo/enterprise#41537

X-original-commit: 53ee18ec0a8d7a84acc3ffe62a8a0c40f893307c
Signed-off-by: Tiffany Chang <tic@odoo.com>
Signed-off-by: Quentin Wolfs (quwo) <quwo@odoo.com>