[FIX] sale_loyalty_taxcloud: order confirmation when 'Lock Confirmed Sales' is enabled
Bug:
When the `sale_loyalty_taxcloud` is installed and 'Lock Confirmed Sales'
is enabled, confirming a SO impossible.

Setup:
- install `sale_management` and `sale_loyalty_taxcloud`
- activate Taxcloud (with test credentials)
- enable 'Lock Confirmed Sales' in the settings

Steps to reproduce:
- create a quotation, fill the necessary fields and add a product
- in the 'Other Info' tab, set the fiscal position to
  'Automatic Tax Mapping (TaxCloud)'
- attempt to confirm the quotation

You should be met with a message stating that you can't modify the tax
on a locked order.

Cause:
This issue was introduced by odoo/enterprise@ea954b8

Community PR: odoo/odoo#120998

opw-3289657

closes odoo/enterprise#40880

Signed-off-by: Nshimiyimana Serge Séna (sesn) <sesn@odoo.com>