[FIX] currency_rate_live: bank of turkey api
Before: trying to update the currencies while selecting the bank oof Turkey as the exchange service would raise a UserError. This was causes by an SSL error that happens with OpelSSL 3.0 when trying to connect to legacy websites that disable renegotiation without signaling it correctly.

Now: an SSL context that enables legacy server connect is set up and used for the server request.

task-3302273

closes odoo/enterprise#43457

X-original-commit: bea8f4d018700907632e3ea393b4c3661b46a8c4
Related: odoo/odoo#126985
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Ali Alfie (alal) <alal@odoo.com>