[FIX] currency_rate_live: bank of turkey api
Before: trying to update the currencies while selecting the bank oof Turkey as the exchange service would raise a UserError. This was causes by an SSL error that happens with OpelSSL 3.0 when trying to connect to legacy websites that disable renegotiation without signaling it correctly.

Now: an SSL context that enables legacy server connect is set up and used for the server request.

task-3302273

closes odoo/enterprise#40936

Related: odoo/odoo#125161
Signed-off-by: Quentin De Paoli <qdp@odoo.com>