[IMP] knowledge: extend embedded view behavior API to support xmlid
To embed a view in an article template, we have to copy the action window
opening that view in the attribute of a HTML tag that will be recognise
by the editor as an embedded view. To reduce the amount of data stored
in the templates and ease their maintenance, we will extend the API of
the embedded view behavior to support xmlid.

task-2796157

Part-of: odoo/enterprise#34663