[FIX] hr_work_contract_attendance: use attendance
Issue:
With a contract that uses attendances,
the work entries generated are always overtimes.

Solution:
Overtimes are only used in the case of contracts
that are based on the calendar.
We must therefore bypass the override
of the `_get_interval_work_entry_type` method
in the case the contract is not based on the calendar.

closes odoo/enterprise#41251

X-original-commit: 127066aaa91451f1bda8cd5483e24fdfba985b67
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>