[FIX] web_studio: tour rename select activity for sure
Before this commit, sometimes but rarely, the test_rename tour selected
account.account to make a many2many relation and enabled "color" on that field.
That color field (or rather, label) conflicted with consolidation_color.

Without being able to reproduce, I have a possible explanation.
In tours, the "text" command first does a click on the target element, then writes text
and then fires input and change events.
In the many2one field widget, the first click triggers a first name_search that, if it returns
in time, may bypass the input and trigger the next step of the tour intempestively. If the next step
doesn't test for the values available in the many2one's dropdown, it way then succeed without having selected
the value we wanted.

Hence this commit, makes the tour wait for a precise name_search value, then clicks on the one we want.
We hope that the undeterministic issue is gone after that.

runbot error: 20240
runbot build: https://runbot.odoo.com/runbot/build/35485652

closes odoo/enterprise#40981

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>