[FIX] documents: fix share upload notification
Documents uploaded by a public user via a share historically had
the record created as if it was done by the user who created the share.

This behavior was kept when `upload_attachment` was modified in
3045852, but the recordset containing all created documents was itself
not linked to the share create_uid. This caused a problem when trying
to post a notification with `message_post`.

Task-3324685

closes odoo/enterprise#41023

X-original-commit: b3fd9e99ff2a195bb1057949c63ba79c07e9840c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>