[IMP] knowledge: add embedded item calendars
Purpose:
--------

Add a new editor command "/item calendar", that allows users to embed a
calendar view showing the items of the current article, using the
properties of these items to display them in the calendar.
Usual calendar features such as creation, edition of "events" (article
items in this case) are working and sets/edits the corresponding item
properties.

This commit introduces an extension of the calendar model, called item
calendar model, that allows to use properties as settings for a calendar
view. To do so, one needs to replace the fieldMappings when initializing
the model in the calendar controller to use the properties.
This requires the fieldMapping to be stored somewhere to work.

This fieldMapping is stored in a new entry of the behaviorProps called
additionalViewProps. This new entry allows to define props to pass to
the embedded view created by the embedded view manager. These props are
also passed to the view created when opening an embedded view in full
screen.
This feature thus only works for embedded views in Knowledge, but one
could extend this feature and store the fieldMapping differently if one
would like to use it somewhere else.

Since the properties are stored in the behaviorProps, these are not
available when opening an items list or kanban. Therefore, the item
calendar view has its own act_window so that one cannot switch from
list/kanban to calendar and vice versa.
The properties are thus also not available when reloading the page when
opened in full screen. In that case, we show an empty calendar.

When inserting the item calendar, a dialog allows the user to select
the properties to use in the item calendar view:
- name (string - name of the view)
- start date (date/datetime property - mandatory)
- end date (date/datetime property - same type as start date
property)
- color (boolean/selection/many2one property - used to show the items
in different groups of colors)
- default scale (calendar scales)
The creation of new date or datetime properties is possible from this
dialog.
If no date/datetime property exists, two default start and stop dates
will be created.

This dialog can also be used to edit an existing embedded view, by
clicking on the edit button of the embedded view manager.

Task-3239985

Part-of: odoo/enterprise#40993