[FIX] knowledge: remove observer lock on Behavior destroy
Ensure that the lock on the editor observer is properly removed when a Behavior
is destroyed, because in some rare cases, a Component can be destroyed before
being mounted or before a patch finish being applied.

Remove redundant `idSet` call `onWillStart` and `onWillPatch` because at that
point, no new content has been applied by OWL, so the Behavior does not have new
elements in the DOM that would not have an editor id.

task-3280704

Part-of: odoo/enterprise#40062
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>