[FIX] knowledge: parallelize instanciation of behaviors
OWL `App.mount` promise is never resolved if the App is destroyed before it is
mounted. This would result in the `updateBehaviors` mutex to be locked forever
on that promise, preventing other new `Behaviors` from being mounted. Using a
Deferred promise that is either resolved when the App is mounted or when it is
destroyed prevents that issue.

This commit also allows the `App.mount` promises for each `Behavior` to be
resolved in parallel during the `updateBehaviors` execution. This will result in
a performance improvement when loading articles with multiple `Behaviors`, since
a sequential update implies that each rendering await a new rendering frame
animation (at least 16ms per Behavior), but a parallel update allows multiple
renderings to be done during the same frame animation.

Following this change, the `setCursor` has to be called at a point where the
Behavior is already mounted.

Remove a redundant `historyStep` call, because it is the `wysiwyg` insert
command which does the step for the Behavior insertion, no need to do it again
when the Behavior is mounted.

task-3280704

Part-of: odoo/enterprise#40062
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>