[FIX] knowledge: wait the updateBehaviors mutex during commitChanges
During the execution of `updateBehaviors`, the contents of `Behaviors` anchors
are removed just before OWL mount the new rendered nodes. If `commitChanges`,
`updateValue` or `_toInline` are called before the contents are mounted, the
value of the html_field is incomplete, therefore they have to wait for
the mutex to be idle, to be sure that every `Behavior` is fully rendered.

In the case of `_toInline`, since it will modify the DOM in place heavily, we
have to destroy all Behaviors to avoid OWL crashing (because it does not support
having its managed nodes being altered by another process). New Behavior
renderings have to be prevented, so the observer and the listener are stopped.

task-3280704

X-original-commit: 30cb08052a4674eaf6fa5258c692dcf1c3c2600b
Part-of: odoo/enterprise#42615
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>