[FIX] knowledge: track parent of destroyed Behavior out of DOM
`destroyBehaviorApp` should ensure that an `OWL App` is destroyed in the DOM. To
do so, it sometimes has to reintroduces nodes in the `behaviorHandler` so that
the destruction actually happens in the DOM. In that case though, it should
ensure that the anchor is properly returned back to its previous parent (that is
not in the DOM) after the App was destroyed, since other processes (like editor
sanitization) could rely on the presence of those nodes (i.e. in a fragment).

task-3280704

Part-of: odoo/enterprise#40062
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>