[FIX] knowledge: test /article command in the Email Composer
Add a test to check that insertion of `/article` works in the Email Composer,
and that two successive `/article` blocks are not merged together on the same
line.

Fix the `openCommandBar` test util so that in case of an `inline` insert (like
a `<a>` tag in a paragraph, the `<br>` element is correctly placed after the
inserted `<a>` element.

task-3280704

Part-of: odoo/enterprise#40062
Co-authored-by: Thomas Josse <thjo@odoo.com>
Co-authored-by: Damien Abeloos <abd@odoo.com>