[IMP] knowledge: introduce knowledge showcase articles
This PR will add article templates that the user can use to quickly
create new articles in Knowledge. This new feature will greatly improve
the onboarding experience of the users on Knowledge: It will showcase
what the user can do with the app and it will help people writing their
first articles by providing pre-configured articles.

The templates can contain all kinds of blocks and can be specific to some
use cases and industries. We can for instance have a template to create
a todo list, a template to do project management, a template to create a
small CRM, a template to load a dashboard, etc.

The user can load a template in two ways:

1. When the article is empty, the interface will show a contextual menu
   in the editor allowing the user to load a template or insert a kanban
   or list view. The contextual menu will disappear as soon as the user
   starts typing something in the editor. When the user clicks on the
   button "Load a Template", a new modal listing all available templates
   will appear. The user can preview the templates and load one inside
   the current article.
2. The Knowledge interface will display a "Browse Templates" button in
   the bottom left corner of the main view. Unlike the "Load a Template"
   button, this button will create a brand new article and load the
   template in it.

This PR will introduce two new models: `knowledge.article.template` and
`knowledge.article.template.category`. The first model will be used to
store the article templates and the configuration that should be applied
to the article (i.e: the property field definition, the icon, etc). The
second model will be used to create categories for the templates. Those
categories will help the users organize their templates.

The templates will have the same hierarchy system as the articles: A
template can have child templates and can not contain cycles. When the
user loads a template, the system will recursively create child articles
from the child templates respecting the same hierarchy.

The templates will currently be stored in the database. By doing that,
people will be able to create their own templates or tweak the default
ones to make sure that they fit their needs. They can for instance edit
the body of the template, add/remove property fields, add/remove
children, etc.

To create a new template:

1. The administrators can go in the "Settings" > "Technical" > "Knowledge"
   > "Templates" and create a new record. The administrators can edit the
   body of the template using a textarea.
2. As the templates are simple records, the templates can be loaded from
   a module using the `data` folder. Those data will be loaded in the
   database when installing the module. People could ship their templates
   in a module and share them with the community.

In the future, we would like to improve the integration of Knowledge with
the Odoo Store to let people install additional article templates in one
click. People could then create modules containing article templates and
sell them online.

We also may want to load a template from an external server to quickly
deliver new templates or update the existing ones without having to
update the Odoo instance of the clients.

task-2796157

Part-of: odoo/enterprise#34663