[IMP] Helpdesk: prevent creation of ghost stages
Purpose: prevent the creation of stages when not viewing the tickets of a specific team. Otherwise, that stage will end up in a sort of limbo: it is not displayed in teams pipelines because it has none set, and it is not visible in cross-teams views if it doesn't have any tickets.
In addition, it can be confusing for users if their newly created stage disappears at the first refresh of the page.

Implementation: fixed by overriding the canCreateGroup in the HelpdeskTicketRenderer, the context of the list now needs
the active_model to be "helpdesk.team", only available from the Helpdesk Team actions

Task-3277199

closes odoo/enterprise#41005

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>