[FIX] hr_payroll_account: don't block journal deletion at uninstall
odoo/enterprise#29020 prevents the unlinking of the
`hr_payroll_account_journal` during uninstallation, however that means
if account itself is not removed (and the entire table dropped)
uninstalling `hr_payroll_account` will leave the journal,
then *reinstalling* will attempt to create a new journal with the same
`code`, which is not valid.

Only block the removal of the journal during normal unlinking, during
uninstallation let it through.

closes odoo/enterprise#40361

Related: odoo/odoo#119808
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>