[FIX] knowledge: allow enter key in embedded views
This commit fixes a bug where the enter key would block the user to use the
quick_create features of kanban views inside an embedded one.

This was caused by a listener linked to every keyboard event that would stop
the propagation of the event no matter what key is used. This was done to
prevent unwanted behaviors coming from the editor such as the `/` key that
would bring up the command popover even though it was not wanted.

We are also bypassing the keydown event listeners of the editor
that prevented default behaviors for the embedded view, such as 'Backspace' to
remove text from an input.
By propagating the events to the parent of the editable we enable the embedded
views to capture the default behaviors for their hotkeys.
This way we let the browser handle the event correctly.

The hotkey service needs the `activeElement` of the ui service to evaluate
potential targets of a hotkey through `getDomRegistrations`.

When the user interacts with the embedded view, we want set the anchor as the
`activeElement` for the embedded view, so that the `querySelector`
of `getDomRegistrations` only captures elements inside of it (we don't want it
to confuse the external view and the embedded view).

The anchor of the embedded view is not editable, so we have to add
`tabindex="-1"` to allow it to be a valid target for `focusin` and `focusout`
events.

We then manually set the `activeElement` as the `anchor` of the embedded view
when it has the focus after `focusin` event, and remove it upon `focusout`, and
at that point, hotkeys used when manipulating the embedded view will trigger
properly without interacting with the containing view (Knowledge Form).

Backport some of 71af7f9 to introduce `setupAnchor` from 16.0 onwards.

task-3302786

closes odoo/enterprise#41008

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>