[FIX] helpdesk: change helpdesk team in ticket mobile view possible
Steps to repoduce:

- Install the helpdesk module.
- Go to helpdesk, go inside any team and select any ticket.
- Now use your browser to inspect and toggle device toolbar (or use a
real smartphone).
- Now inside the ticket view form, try to change the ticket helpdesk
team.

Issue:

We can't change the helpdesk team as we don't get a proper view adapted
to mobile view that allows us to easily change the team.

Solution:

Added proper context with `kanban_view_ref` to ref to the proper view
for the helpdesk team. This fix is similar to:
https://github.com/odoo/enterprise/pull/39499/

opw-3282805

closes odoo/enterprise#41012

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>