[FIX] appointment: fix some `.btn-secondary` border
== ISSUE ==

With the new design introduced within Milk, most of the buttons received
a visual upgrade (eg. removing borders)

This change applied on most of the buttons since the same code, but some
buttons were using a `.border` class on top of that and so they did not
loose this border.

== After this commit ==

We clean up these buttons that were keeping a border, in order to make
our secondary buttons more consistent.

Please note that this change does not apply on all secondary buttons
since some of them are customized for a specific module and so remain
unmodified.

task-3326313
part of task-3326263

closes odoo/enterprise#41310

X-original-commit: bbae33e6f8fa6636c59005eb8b74c4be194ab400
Related: odoo/odoo#121990
Signed-off-by: Pierre Paridans (app) <app@odoo.com>