[FIX] web_map: text contrast on buttons
== ISSUE ==

If you go to the Contact module and open the map view, primary buttons
also have a `.text-white` class which will apply a weird background
when clicking on the buttons. This makes its content not readable.

== After this commit ==

We remove the class from the button since it's not needed and causes an
issue in terms of accessibility.

task-3326312
part of task-3326263

closes odoo/enterprise#41249

X-original-commit: 5d2ed3915e98978c7de63f931ff54c469841da30
Signed-off-by: Pierre Paridans (app) <app@odoo.com>