[IMP] appointment: invite colleagues to meeting
Purpose
========
Allow people to invite colleagues to the meeting they're registering too.
e.g. : As a CEO scheduling an Odoo demo, I want to easily add my CTO to the list
of attendees as I want them to be there too.

Specifications
===============
1.Add a boolean field called "Allow Guests" with helper "Let attendees invite
guests while registering to this meeting" on the options page of appointment.
2.If boolean field is checked then a linked-style "+Add Guests" button is added,
in the appointment form.
3.When clicked, the link become text and a text-area is appear with a close
button on the top right of the textarea.
4.In the text area one can write email address line by line and from that email,
attendees is added in the event.
5.On the appointment confirmation page display all the attendees and "+add guest
button" to add the guest again if user wants.

Technical
==========
1.If user write a email address which is not a partner then the partner is
created and name of the partner is set as email address.(this is done on both
appointment form and confirmation page).
2.On the appointment form page if a partner is unavailable then he is redirected
to the calendar page to change the date but on the confirmation page,if partner
is unavailable then an error message is shown that guest is unavailable .

Task-3297751

closes odoo/enterprise#45332

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>