[FIX] knowledge: fix invisible article body on mobile
When the user is on a mobile device, the body of the article is no longer
visible. The reason is that the main container of the article no longer
have a `flex` display on small screen size. As a result, the children of
that container do not stretch to take the full height of the view. The
scroll view containing the body of the article will then have a height
of 0 which makes it invisible. This commit will fix that issue by
replacing the class `d-sm-flex` to `d-flex`.

Related: https://github.com/odoo/enterprise/commit/7fc9528e30aff39a3a0261fc81504b726bc8f98a
task-3326424

closes odoo/enterprise#41218

X-original-commit: e728863142e0e2843c0a9a8252f698615d2624ef
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>