[FIX] sign: color fixes
There were still old, hardcoded, colors being used in the sign module's
PDF viewer.
These have been replaced with the new ones, but Sign uses an
iframe to render the PDF viewer and the variables weren't being applied.
Some of the bootstrap classes, such as `.btn-primary` (changed to
`.btn-secondary`), were already on some elements but were also not being
applied.
A copy of the current CSS variables was needed in the `iframe.scss` file
in order to fix these issues.

This also allows us to correct some of the viewer's color issues, which
we override in this same file.

Inputs with the class `o_input` just above the iframe weren't rendering
well as they were white instead of being the same as the parent's
background color. To solve this issue we override the value of the CSS
variable being used by `.o_input`.

task-3326565
part of task-3326263

Forward-port-of: #41035
X-original-commit: 0804065d9e9190002125c2686435fac8f46b99ec
Part-of: odoo/enterprise#44948