[FIX] web_enterprise: rounded custom app icons
=== ISSUE ===

If you activate studio and design a custom icon, it does not have any
`border-radius`. This is due to the fact that the icon is inside a `<div>`
and not inside an `<img>`.

=== AFTER ===

We had a `rounded-3` class to the div containing the icon. This value
comes from the one applied on the regular app icons.

task-3326580
part of task-3326263

closes odoo/enterprise#41297

X-original-commit: d52bb8cf5b4f295b5d7e5ad6c9b08f06dd4aeb0f
Signed-off-by: Pierre Paridans (app) <app@odoo.com>