[FIX] l10n_co_reports: fix certification report PDF
Steps to reproduce:
- Install l10n_co_reports and switch to the Colobian company.
- Create and post a vendor bill for a partner using an account
that starts with 2365 (but different from 236505).
- Repeat for a second partner.
- Open the 'Certificado de Retención en Fuente'.
- Leave all partners folded and click on 'PDF'.

Observed behavior:
The PDF contains all the partners.

Desired behavior:
The file should only contain the unfolded partners.

This is because Colombian exports do not use the generic
`export_to_pdf()` method, which in trun calls
`_filter_out_folded_children()` on the results of the
`_get_lines()`.
The solution is simply to manually call
`_filter_out_folded_children()` in the Colombian export
method.

closes odoo/enterprise#41104

X-original-commit: 7e23d1f174594d629bf699781ca954334b14cfd5
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>