[FIX] sign: fix rotateCw arrow display
The PDF viewer used to be black. As it is now white, and the custom
rotateCw arrow that is injected is a white `.png` image, it is no longer
visible on the white background. There is also a sizing issue.
The fix was to revert to the original rotateCw icon located in the PDF
viewer and use the same way of working as for the other
buttons. To do this a simple ID was needed on the element.

task-3326565
part of task-3326263

Forward-port-of: #41035
X-original-commit: 3f2c40bd23c12c6802f6aedbcc1b99239b15de9c
Part-of: odoo/enterprise#44948