[FIX] data_merge: limit company_id search to own model
The subquery to search for records having a certain company id
uses a left join. When the where clause includes 'company_id
IS NULL' the join will keep all the 'data_merge_record' records
even if they have a different res_model_id than the model we
are joining with.
To limit the results to the target model in each subquery,
we can add the res_model_id again in the where clause and
filter out the wrong records.

closes odoo/enterprise#41740

X-original-commit: fa1345f3eec554bc0f602888fd05dc9ab79d664a
Signed-off-by: Kevin Baptiste <kba@odoo.com>