[FIX] partner_commission: order by sequence
Steps to reproduce:
- install sale_management, crm and partner_commission;
- create a commission plan with two rules for the
same category of product but one with a product and one without;
- link this commission plan and a partner level
to a partner in "Partner Assignment" tab;
- create an order with the product specified in the rule.

Issue:
The commission plan applied is not the right one.

Cause:
To find the commission plan rule to apply,
we order all possible rules with the following
order: `product_id`, `template_id`, `pricelist_id`.
Then we take the first rule.
The desired result is that if there is
a `product_id` defined, it is at the top of the list.
Unfortunately, to sort the `product_id`,
we use the following order:
`'priority desc, default_code, name, id'`.
The null ids will not necessarily be in
the last position and vice versa.

Solution:
Order the rules by sequence.
This solution is intuitive.
With all the rules that match,
we take the one we have defined as the first.

opw-3285638

closes odoo/enterprise#41152

X-original-commit: 202af610f5bd6ec41f1f2a98f9fd4f14cd810a53
Signed-off-by: Arnaud Joset <arj@odoo.com>
Signed-off-by: Lefebvre Thomas (thle) <thle@odoo.com>