[FIX] web_studio: <setting> should not crash
Have a <setting> inside a innergroup
```xml
<form><group><group>
    <setting .... />
</group></group></form>

```

Before this commit, opening that form view in studio triggered a crash because of props
validation. The setting component received the studioXpath props, but was not set for it

Now, the setting component in studio accepts that props, yet it is not used. There is no crash anymore.
Note that there is no handler to customize settings, so that node in studio is inert.

task-3305800

closes odoo/enterprise#41080

X-original-commit: 4d7b313
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>