[IMP] web_enterprise, *: app icons: replace svg to png
The introduction of Milk brought a new panel of app icons.
Using the svg format creates a lack of anti-aliasing on the edges of the
shapes, which makes the icons look bad.
Since the png size has been reduced, we can afford to use the png format
to have the best possible quality without having a lack of performance.

task-3326633
Part of task-3326263

X-original-commit: 1a7902c9e52bae1662de0508f16f005d7ad78423
Part-of: odoo/enterprise#41279