[FIX] l10n_de_pos_cert: escape line breaks in product name
Steps to reproduce:

Make a product with a line break (e.g. `\n`) in its name. add it to
a PoS order, validate and attempt to close the session.

Get an error: `SyntaxError: EOL while scanning string literal`

The problem is that line breaks are not escaped, and cause the
while rendering the `dsfinvk_cash_point_closing_template` view.

opw-3265441

closes odoo/enterprise#41128

X-original-commit: a85abbab8932bb4dbb4bb3a69824377d31e3c665
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Khanalizadeh Ahmad (khah) <khah@odoo.com>