[IMP] industry_fsm_sale: make log on generated SO for task
Prior to this commit, when creating a new quotation from a field service task,
a link to the generated quote was available from the task. However, there was
no corresponding link from the quote back to the field service task.

With this commit, a log entry will be added to the generated sales order,
containing the name of the task. The task name in the log entry will be
clickable, allowing users to conveniently redirect to the corresponding task record.

task-3251766

Part-of: odoo/enterprise#41569