[FIX] sale_stock_renting : draft after cancel rental orders
When we create a rental order with a service which have plan services and confirm it ,
then cancel it and then set to quotation (set to draft again) it will show an error.

Steps to reproduce the error :
1- install rental and planning
2- create a product set its type as service , and select plan services , select developer for example
3- create a rental order with that product and confirm it
4- Cancel the created order and the click on "set to quotation"
5- You will get an error

The error was coming from the a cache miss for the value virtual_available_at_date (https://github.com/odoo/odoo/blob/6b7f3f6ec40eb819dd1d94a94722610a63d5697a/odoo/fields.py#L1214C1-L1215) .
On the update I checked that the value is indeed in the cache but setting up warehouse_id to False seems to be the source of the problem since the
function depends on it and we update it later to False.

opw-3300970

closes odoo/enterprise#41367

X-original-commit: 0b9e541a77896fcd3db95cfd55a714a5d77cd100
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>