[FIX] stock_barcode: mute sound during tests
We add a error sound in the Barcode App in odoo/enterprise#33383
The problem is, while running locally the test tours, even if chromium
is run headless, the sounds are played. Those sounds coming from nowhere
can drive anyone insane !
To fix that, this commit adds a new config parameter,
`stock_barcode.mute_sound_notifications`. If set, it will disable the
sound in the Barcode App (and it is in the tests' setup !).

Also, moves the sound related code from `BarcodeModel` to
`MainComponent` so it's easier to manage that as a config and it seems
more logic for the sounds to be managed by a Component ("client side"
object) instead of by the `BarcodeModel (data manipulating object).

closes odoo/enterprise#41426

X-original-commit: a747623c4a3e9d9367870aef33399251b7bacc56
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>