[FIX] account_invoice_extract: fix nondeterministic failure in tests
When the `test_no_merge_check_status` is ran at around midnight, it
could happen that it failed because of the `invoice_due_date` not being
set to the expected value.

This was due to the fact that we rely on the `create_date` and
`invoice_due_date` to decide if the due date should be filled or not.

Unfortunately, when the `invoice_due_date` is set to "now" there can be
a few seconds of difference with the `create_date` leading to a mismatch
when one is set before midnight and the other after midnight.

The test will now force that `invoice_date_due` is set to exactly
`create_date`.

closes odoo/enterprise#41465

X-original-commit: 7586f49ff92c1b80d8e92d4c36645f2a267ee53c
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>