[IMP] {website_}documents: allow to choose domain of share link
In some cases, it might be useful to be able to choose which domain you want to
use to share a workspace or file(s) in Documents. For instance, you might have
two eshops with high-ends products with big margins on one hand and low-end
products with low-margin on the other. For obvious reasons, you don't want
people to see that as they could complain about it.

For those reasons, when sharing documents or a folder, the user can now choose
the domain of the shared URL by selecting on which website the link is shared.

By default, the website of the company folder is selected. If there is no
company on that folder or the company has no website, the current company
website is selected (which might be "False"). The user can override the default
choice.

The base URL is determined by the website associated with the sharing. If the
website is set and has a "domain", it is used as base URL. If not, the
web.base.url of the config parameters is used.

The shared URL is automatically updated when the website is changed. When
deleting a website, all sharings associated with it are deleted as well.

Note that no matter the website chosen, the shared documents/folder can be
accessed with any domain configured on the server instance.

Task-3325062

closes odoo/enterprise#42024

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>