[FIX] account_reports: wrong test
When reversing an account.move of type 'entry' containing taxes, the same repartition must be used as on the original move, so that everything is cancelled in the tax report (unlike invoices, which will switch to the refund repartition). This test was instead checking the result in the tax report corresponded to the refund repartition ... Though the test was only passing by chance, because of the bugs solved in https://github.com/odoo/odoo/pull/121605. The repartition that was used actually was the invoice repartition, not the refund one, but the wrong values computed for tax_tag_invert caused the reverse to still appear in positive in the report. The confusion came from the fact the refund repartition impacts the same tags as the invoice one on the test tax used here, with the same signs.

closes odoo/enterprise#41162

Related: odoo/odoo#121605
Signed-off-by: Quentin De Paoli <qdp@odoo.com>