[IMP] event_barcode: add qr code on tickets
As qr codes are simpler to scan on smartphones, they
are added alongside the existing barcode on event full
page tickets as well as foldable badges. Space on tickets
is reworked and redistributed to do so. The example
tickets on events are also given a default qr code.
The qr code is based on the barcode and represents
the same registration number.

The route /report/barcode is used to generate the codes.
It allows users to access their tickets even if they are
not logged in.

UPG PR - odoo/upgrade#5007
COM PR - odoo/odoo#120933

Task-3061851

closes odoo/enterprise#41164

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>