[FIX] sale_subscription: Improve perf by adding missing indexes
Current behavior:
In some instances, like when the customer is on a shared instance worker the query execution time is suboptimal resulting in significant delays in loading the KPI report in the subscription module. Missing indexes in standard fields primarily caused the issue and the query takes around 50 seconds to execute and sometimes it results in a deadlock.

Error:
Slow down in loading the KPI report.

Expected behavior:
Open KPI report with no problem.

Fix:
To address this issue, we added the necessary indexes and as a result, the query execution time improved significantly, reducing it from approximately 50 seconds to just 100 milliseconds.

Affected versions:
* > 14

opw-3249090

closes odoo/enterprise#42135

X-original-commit: 68d72007329861cf2e2894084a917ea028b28386
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Noura El Allam (nea) <nea@odoo.com>