[FIX] appointment: prevent traceback when email template is not found
When a user deletes the email template `appointment_booked_mail_template` which
is present in the data file of the `Appointments` module. Then the user creates
a new meeting in online appointments and saves it at that time traceback is
generated on the user side and also in the log.

Steps to reproduce :
1. Install the appointment module.
2. Open settings > companies > Review All Templates.
3. Delete the `Appointment: Appointment Booked` email template.
4. Open the calendar app.
5. Open an online appointment and select any one appointment.
6. Click on the appointment state button.
7. Open the list view and create one meeting and save it.
8. Traceback generated.

Applying this commit will fix this issue.

sentry - 4180359024

closes odoo/enterprise#42158

X-original-commit: 45ad86018b4610ddc53f560627ef7672ec16b81a
Signed-off-by: Hennecart Jérémy (jeh) <jeh@odoo.com>