[FIX] web_enterprise: home_menu: truncated labels
Prior to this commit, the introduction of Milk allowed the app switcher
to be on two lines. But they were not truncated when they were composed
of a long word.

To fix this, this commit reverts to a one-line label that can be
truncated if necessary.

task-3329928
Part of task-3326263

closes odoo/enterprise#41298

X-original-commit: 5b04371dbbdcc9394d482c84db819292299c397c
Signed-off-by: Pierre Paridans (app) <app@odoo.com>